+++ src/lib/simpleio/control.c
@@ -18,9 +18,9 @@ con_cmd con_commands[] = {
     { 2, cmd_goto },    // 1
     { 1, cmd_clr },     // 2
     { 1, cmd_set },     // 3
-    { 0, cmd_getx },    // 4
-    { 0, cmd_gety },    // 5
-    { 0, cmd_null },    // 6
+    { 1, cmd_get },     // 4
+    { 0, cmd_getx },    // 5
+    { 0, cmd_gety },    // 6
     { 0, cmd_null },    // 7
     { 0, cmd_null },    // 8
     { 0, cmd_null },    // 9