+++ src/lib/simpleio/control.h
@@ -5,8 +5,9 @@
 #define TERM_CMD_GOTO    1  // Position cursor.
 #define TERM_CMD_CLR     2  // Clear flags.
 #define TERM_CMD_SET     3  // Set flags.
-#define TERM_CMD_GETX    4  // Get X position.
-#define TERM_CMD_GETY    5  // Get Y position.
+#define TERM_CMD_GET     4  // Set flags.
+#define TERM_CMD_GETX    5  // Get X position.
+#define TERM_CMD_GETY    6  // Get Y position.
 
 #define TERM_FLAG_CURSOR   1    // Cursor visibility.
 #define TERM_FLAG_REVERSE  2    // Reverse mode.
@@ -17,6 +18,7 @@ extern void cmd_null (void);
 extern void cmd_goto (void);
 extern void cmd_clr (void);
 extern void cmd_set (void);
+extern void cmd_get (void);
 extern void cmd_getx (void);
 extern void cmd_gety (void);
 extern void cmd_lf (void);